Correct use of conditional statements
Hi Matt,

See my suggested changes to your file. Note that when you place a semi-colon at the end of a conditional statement, it "short-circuits" the conditional statement. Every line in your code should end with a semi-colon – except the start of a conditional statement or a function.

On a different note, you have written code that is well organized (with good use of whitespace to separate blocks of code) and you have commented your code well to describe intent.

Mr. Gordon