Suggestions for readability and efficiency
Hi Adam,

You've clearly put some effort into improving the stars effect you started last class! I can't wait to see what it looks like when run.

Please have a look at my proposed changes – you've largely made good use of whitespace in your code to separate logical blocks of code, but could use more comments in places to describe intent.

For the if statement – instead of having three separate if statement – I've combined it into a single if statement with two else-if branches. This is more efficient – the computer will only have to check one or two of the branches of the if statement (depending on the value of the scene variable) rather than always running three separate IF statements. It's a small thing, but we strive for efficiency in our code. 

You've clearly put effort into your work – keep it up!

See you in class,

Mr. Gordon